requirements.txt
torch==2.2.2
torchvision==0.17.2
gradio==4.44.0
numpy==1.26.4
pillow==10.4.0
matplotlib==3.9.2
scikit-learn==1.5.1
tqdm==4.66.5
